refactor: migrate baseComputedStatsObject
