+++ src/lib/gpu/injection/injectPrecomputedStats.ts
@@ -1,6 +1,6 @@
-import { baseComputedStatsObject } from 'lib/conditionals/conditionalConstants'
 import { GpuConstants } from 'lib/gpu/webgpuTypes'
 import { ComputedStatsArray, Key, KeysType } from 'lib/optimization/computedStatsArray'
+import { baseComputedStatsObject } from 'lib/optimization/config/computedStatsConfig'
 
 export function injectPrecomputedStatsContext(x: ComputedStatsArray, gpuParams: GpuConstants) {
   const a = x.a
