+++ src/lib/gpu/webgpuDebugger.ts
@@ -1,7 +1,7 @@
-import { baseComputedStatsObject } from 'lib/conditionals/conditionalConstants'
 import { Sets } from 'lib/constants/constants'
 import { GpuExecutionContext } from 'lib/gpu/webgpuTypes'
 import { ComputedStatsArray, ComputedStatsArrayCore, ComputedStatsObjectExternal, InternalKeyToExternal, Key, KeysType } from 'lib/optimization/computedStatsArray'
+import { baseComputedStatsObject } from 'lib/optimization/config/computedStatsConfig'
 import { TsUtils } from 'lib/utils/TsUtils'
 
 export function logIterationTimer(i: number, gpuContext: GpuExecutionContext) {
@@ -21,29 +21,29 @@ export function debugWebgpuOutput(gpuContext: GpuExecutionContext, arrayBuffer: 
 }
 
 /*
-    ED: array[22],
-    BASIC: array[69],
-    SKILL: array[70],
-    ULT: array[71],
-    FUA: array[72],
-    DOT: array[73],
-    BREAK: array[74],
-    COMBO: array[75],
-    EHP: array[77],
-    HEAL: array[102],
-    SHIELD: array[105],
-    xHP: array[4],
-    xATK: array[5],
-    xDEF: array[6],
-    xSPD: array[7],
-    xCR: array[8],
-    xCD: array[9],
-    xEHR: array[10],
-    xRES: array[11],
-    xBE: array[12],
-    xERR: array[13],
-    xOHB: array[14],
-    xELEMENTAL_DMG: array[22],
+ ED: array[22],
+ BASIC: array[69],
+ SKILL: array[70],
+ ULT: array[71],
+ FUA: array[72],
+ DOT: array[73],
+ BREAK: array[74],
+ COMBO: array[75],
+ EHP: array[77],
+ HEAL: array[102],
+ SHIELD: array[105],
+ xHP: array[4],
+ xATK: array[5],
+ xDEF: array[6],
+ xSPD: array[7],
+ xCR: array[8],
+ xCD: array[9],
+ xEHR: array[10],
+ xRES: array[11],
+ xBE: array[12],
+ xERR: array[13],
+ xOHB: array[14],
+ xELEMENTAL_DMG: array[22],
  */
 
 export function debugExportWebgpuResult(array: Float32Array) {
